requirements.txt
imgaug
imageio
imageio-ffmpeg
opencv-python
typing>=3.7
Click>=7.0
numpy>=1.17
torchsummary>=1.5
matplotlib>=3.1
termcolor>=1.1
torchvision>=0.2
Pillow>=6.2
tensorboardX>=1.9
PyYAML>=5.1.2
path.py>=12.0
scipy>=1.6.3
pycocotools>=2.0.2